requirements.txt
pandas==1.3.4
pyyaml==5.4.1
yfinance==0.1.63